backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]==3.2.13
gunicorn==23.0.0
django-filter==24.3
neo4j==5.25.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.1.1
channels==4.1.0
channels-redi